requirements.txt
giskard~=2.1.1
giskard[llm]~=2.1.1
numpy~=1.26.2
pandas~=2.1.4
scipy~=1.11.4
datasets~=2.15.0
transformers~=4.36.1
requests~=2.31.0
tenacity
openai 
unstructured 
pdf2image 
pdfminer-six 
faiss-cpu
langchain
pypdf
tiktoken
python-dotenv